+++ blockchain/block.py
@@ -52,17 +52,12 @@ class Block:
         return json.dumps(json_dictionary)
 
 
-def block_to_string(json_block):
-    index = int(json_block['index'])
-    current_hash = json_block['hash']
-    prev_hash = json_block['prev_hash']
-    data = json_block['data']
-    nonce = int(json_block['nonce'])
-    generated_node_number = json_block['This block generated by Node '][0]
-
-    return f'This block generated by Node [{generated_node_number}]:     Index = {index}, Hash = {current_hash}, ' \
-           f'Prev_Hash = {prev_hash}, Data = {data}, Nonce = {nonce}'
+# Генерация генезиса
+def create_genesis():
+    genesis_block = create_new_block(0, 'GENESIS', 1, -1)
+    return genesis_block.block_to_json()
 
 
+# Генерация нового блока
 def create_new_block(index, _prev_hash, nonce_type, server_id):
     return Block(index, _prev_hash, nonce_type, server_id)
