+++ blockchain/node.py
@@ -1,11 +1,4 @@
-import block
 import json
-
-
-# Генерация генезиса
-def create_genesis():
-    genesis_block = block.create_new_block(0, 'GENESIS', 1, -1)
-    return genesis_block.block_to_json()
 
 
 class Node:
@@ -26,7 +19,7 @@ class Node:
         if received_index == 0:
             self.blocks_array.append(received_block)
             self.block_index = 0
-            print(block.block_to_string(python_object))
+            print(block_to_string(python_object))
             return True
 
         # Это блок, которого у нас еще нет -> запишем его
@@ -36,7 +29,19 @@ class Node:
             self.blocks_array.append(received_block)
             self.block_index = received_index
             if self.server_id != python_object['This block generated by Node '][0]:
-                print(block.block_to_string(python_object))
+                print(block_to_string(python_object))
             return True
 
         return False
+
+
+def block_to_string(json_block):
+    index = int(json_block['index'])
+    current_hash = json_block['hash']
+    prev_hash = json_block['prev_hash']
+    data = json_block['data']
+    nonce = int(json_block['nonce'])
+    generated_node_number = json_block['This block generated by Node '][0]
+
+    return f'This block generated by Node [{generated_node_number}]:     Index = {index}, Hash = {current_hash}, ' \
+           f'Prev_Hash = {prev_hash}, Data = {data}, Nonce = {nonce}'
