+++ tests/block_test.py
@@ -98,33 +98,22 @@ def test_block_to_json():
         assert nonce == new_block.nonce
 
 
-def test_block_to_string():
+def test_create_genesis():
     for i in range(100):
-        block_index = random.randint(1, 100000)
-        nonce_type = random.randint(1, 3)
-        prev_hash = 'This is old String block'
-        server_id = random.randint(1, 3)
+        genesis_block = block.create_genesis()
 
-        new_block = block.create_new_block(block_index, prev_hash, nonce_type, server_id)
+        assert genesis_block is not None
+        assert type(genesis_block) == str
 
-        json_block = new_block.block_to_json()
-        assert type(json_block) == str
+        python_object = json.loads(genesis_block)
+        genesis_generated_by = python_object['This block generated by Node '][0]
+        genesis_index = int(python_object['index'])
+        genesis_hash = python_object['hash']
+        genesis_prev_hash = python_object['prev_hash']
+        genesis_data = python_object['data']
 
-        python_object = json.loads(json_block)
-
-        string_block = block.block_to_string(python_object)
-        assert type(string_block) == str
-
-        generated_by = int(string_block.split('This block generated by Node [')[1].split(']:')[0])
-        index = int(string_block.split('Index = ')[1].split(', Hash')[0])
-        cur_hash = string_block.split(', Hash = ')[1].split(', Prev_Hash')[0]
-        prev_hash = string_block.split('Prev_Hash = ')[1].split(', Data')[0]
-        data = string_block.split('Data = ')[1].split(', Nonce')[0]
-        nonce = int(string_block.split('Nonce = ')[1])
-
-        assert generated_by == new_block.server_id[0]
-        assert index == new_block.index
-        assert cur_hash == new_block.hash
-        assert prev_hash == new_block.prev_hash
-        assert data == new_block.data
-        assert nonce == new_block.nonce
+        assert genesis_generated_by == -1
+        assert genesis_index == 0
+        assert genesis_hash[-4:] == "0000"
+        assert genesis_prev_hash == 'GENESIS'
+        assert len(genesis_data) == 256
